requirements.txt
# 核心依赖
aiohttp>=3.9.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# 可选依赖
# websocket-client>=1.6.0
# pandas>=2.0.0
